Improve robustness of student view
