+++ frog/imports/ui/StudentView/Runner.jsx
@@ -8,6 +8,8 @@ import { activityTypesObj } from '../../activityTypes';
 import { ActivityData, reactiveFn } from '../../api/activityData';
 import { createLogger } from '../../api/logs';
 import { saveProduct } from '../../api/products';
+import { Objects } from '../../api/objects';
+import { Activities } from '../../api/activities';
 
 const Runner = ({ activity, object, reactiveKey, reactiveList }) => {
   const activityType = activityTypesObj[activity.activityType];
@@ -32,22 +34,28 @@ const Runner = ({ activity, object, reactiveKey, reactiveList }) => {
 };
 
 export default createContainer(
-  props => {
+  ({ activityId }) => {
     // there is one reactiveKey Object per groupId
     const reactiveKey = ActivityData
       .find({
-        activityId: props.activity._id,
+        activityId,
         type: 'kv'
       })
       .fetch();
 
     const reactiveList = ActivityData
       .find({
-        activityId: props.activity._id,
+        activityId,
         type: 'list'
       })
       .fetch();
-    return { ...props, reactiveKey, reactiveList };
+
+    const o = Objects.findOne({ activityId });
+    const object = o ? o.data : null;
+
+    const activity = Activities.findOne({ _id: activityId });
+
+    return { activity, object, reactiveKey, reactiveList };
   },
-  props => <Runner {...props} />
+  Runner
 );
